menu items moved to own dir.  adding menuBar using pm example
